requirements.txt
pydantic
pyyaml
azure-storage-blob
azure-data-tables
azure-servicebus
azure-core
python-dotenv
